Se realizan correcciones, sobre observaciones echas en pr: limpieza de codigo
